// src/interfaces/Pokemon.tsx
export interface Pokemon {
  name: string;
  url:  string;
  id: number | undefined;
  details : PokemonDetails | undefined;
}


export interface PokemonDetails {
  abilities:                Ability[];
  base_experience?:          number;
  forms?:                    Species[];
  game_indices?:             GameIndex[];
  height?:                   number;
  held_items?:               any[];
  id?:                       number;
  is_default?:               boolean;
  location_area_encounters?: string;
  moves?:                    Move[];
  name:                     string;
  order?:                    number;
  past_types?:               any[];
  species?:                  Species;
  sprites?:                  Sprites;
  stats?:                    Stat[];
  types?:                    Type[];
  weight?:                   number;
}

export interface Ability {
  ability:   Species;
  is_hidden: boolean;
  slot:      number;
}

export interface Species {
  name: string;
  url:  string;
}

export interface GameIndex {
  game_index: number;
  version:    Species;
}

export interface Move {
  move:                  Species;
  version_group_details: VersionGroupDetail[];
}
// ...
export interface Sprites {
  back_default?:       string;
  back_female?:        null;
  back_shiny?:         string;
  back_shiny_female?:  null;
  front_default:      string;
  front_female?:       null;
  front_shiny?:        string;
  front_shiny_female?: null;
  other?:             Other;
  versions?:          Versions;
  animated?:          Sprites;
}
// ...
export interface Stat {
  base_stat: number;
  effort:    number;
  stat:      Species;
}

export interface Type {
  slot: number;
  type: Species;
}


export interface PokemonForm {
  name: string,
  id: number,
  image : string,
  type: string,
  abilities: string
}